add null checking to book repo delete method

+++ hw06/src/main/java/ru/otus/hw/repositories/BookRepositoryJpa.java
@@ -40,6 +40,8 @@ public class BookRepositoryJpa implements BookRepository {
     @Override
     public void deleteById(long id) {
         Book book = em.find(Book.class, id);
-        em.remove(book);
+        if (book != null) {
+            em.remove(book);
+        }
     }
 }
